Move kernel to _core
